+++ crates/fraud_proofs/global_merkle_root/storage/src/structured_storage/contracts.rs
@@ -1,0 +1,20 @@
+use crate::{
+    column::TableColumn,
+    merkle::MerkleizedTableColumn,
+};
+use fuel_core_storage::tables::{
+    ContractsLatestUtxo,
+    ContractsRawCode,
+};
+
+impl MerkleizedTableColumn for ContractsRawCode {
+    fn table_column() -> TableColumn {
+        TableColumn::ContractsRawCode
+    }
+}
+
+impl MerkleizedTableColumn for ContractsLatestUtxo {
+    fn table_column() -> TableColumn {
+        TableColumn::ContractsLatestUtxo
+    }
+}